+++ frameworks/Rust/actix/Cargo.toml
@@ -1,18 +1,18 @@
 [package]
 name = "actix"
-version = "0.4.1"
+version = "0.5.0"
 
 [[bin]]
 name = "actix"
-path = "src/simple.rs"
-
-[[bin]]
-name = "actix-diesel"
 path = "src/main.rs"
 
 [[bin]]
+name = "actix-diesel"
+path = "src/main_diesel.rs"
+
+[[bin]]
 name = "actix-pg"
-path = "src/pg.rs"
+path = "src/main_pg.rs"
 
 [dependencies]
 askama = "0.6"
@@ -31,7 +31,7 @@ postgres = "0.15"
 horrorshow = "0.6"
 
 actix = "0.5"
-actix-web = "^0.5"
+actix-web = "^0.6"
 
 [profile.release]
 lto = true
